Fixed camera not accounting for scaling

+++ src/Engine/Engine.hpp
@@ -212,8 +212,8 @@ private:
 
 	// Canvas
 	RenderTexture2D m_canvas;
-	float m_virtualWidth = 0;
-	float m_virtualHeight = 0;
+	u32 m_virtualWidth = 0;
+	u32 m_virtualHeight = 0;
 
 	bool m_running = true;
 };
